requirements.txt
aws-cdk-lib>=2.0.0
constructs>=10.0.0
python-dotenv>=0.19.0
PyGithub>=1.55
requests>=2.26.0
pandas>=1.3.0
